include glm in cmake, fixed few warnings

+++ src/main.cpp
@@ -29,7 +29,7 @@ int main()
             if (event.type == sf::Event::Resized)
             {
                 // update the view to the new size of the window
-                sf::FloatRect visibleArea(0, 0, event.size.width, event.size.height);
+                sf::FloatRect visibleArea(0.0f, 0.0f, static_cast<float>(event.size.width), static_cast<float>(event.size.height));
                 window.setView(sf::View(visibleArea));
             }
         }
